reduce login diff
